Add files via upload
Exploring Mushrooms
When asked to think of a mushroom, you might imagine the following:

this shows a photo of some generic button mushrooms in a forest.

It’s a beige, convex mushroom top with a uniform, thick stem. Maybe you thought of it cut up on a slice of delicious pizza or braised with sauce over rice. Regardless, you most likely did not consider this:

A photo of some beautiful mushrooms in a forest. They do not have the signature "mushroom top" and have visually appealing grooves that make them look almost floral.

Mushrooms exist in a variety of different colors, shapes, sizes, textures, etc. In this project, you will analyze an extensive mushroom dataset from UCI using bar charts and acquaint yourself with the diverse array of mushrooms that exist worldwide.